+++ lib/common.py
@@ -1,3 +1,6 @@
+# (c) TenserTensor || Apache-2.0 (apache.org/licenses/LICENSE-2.0)
+
+
 import comfy.samplers as S